+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # Rurik Changelog
 
+## 2024-08-27 - v0.5.1.5 - AI improvements
+* AI can complete some more deed cards - 18 of 32
+* bug: warfare track goes beyond 10.
+
 ## 2024-08-25 - v0.5.1.4 - Bug fixes
 * Fix case at end of round where AI would not progress the game.
 * Don't allow deed cards to be accomplished more than once.
 
 ## 2024-08-25 - v0.5.1.3 - AI improvements
-* AI can complete some deed cards
+* AI can complete some deed cards - 12 of 32
 * bug: Clear the Game Log with "Leave Game"
 
 ## 2024-08-24 - v0.5.1.2 - package binary.